Conver the pages to typescript
