--- pages/api/hello.js
@@ -1,3 +1,0 @@
-export default (req, res) => {
-  res.status(200).json({ message: "Hello world" });
-};
