+++ pages/index.tsx
@@ -1,0 +1,54 @@
+import { GetStaticProps } from "next";
+import Head from "next/head";
+import Link from "next/link";
+import Date from "../components/date";
+import Layout, { siteTitle } from "../components/layout";
+import { getSortedPostsData } from "../lib/posts";
+import utilStyles from "../styles/utils.module.css";
+
+interface Props {
+  allPostsData: { id: string; title: string; date: string }[];
+}
+
+const Home: React.FC<Props> = ({ allPostsData }) => (
+  <Layout home>
+    <Head>
+      <title>{siteTitle}</title>
+    </Head>
+
+    <section className={utilStyles.headingMd}>
+      <p>[Your self introduction]</p>
+
+      <p>
+        (This is a sample website - you'll be building a site like this on{" "}
+        <a href="https://nextjs.org/learn">our Next.js tutorial</a>)
+      </p>
+    </section>
+
+    <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
+      <h2 className={utilStyles.headingLg}>Blog</h2>
+      <ul className={utilStyles.list}>
+        {allPostsData.map(({ id, date, title }) => (
+          <li key={id} className={utilStyles.listItem}>
+            <Link href={`posts/${id}`}>
+              <a>{title}</a>
+            </Link>
+
+            <br />
+
+            <small className={utilStyles.lightText}>
+              <Date dateString={date} />
+            </small>
+          </li>
+        ))}
+      </ul>
+    </section>
+  </Layout>
+);
+
+export const getStaticProps: GetStaticProps = async () => {
+  const allPostsData = getSortedPostsData();
+  return { props: { allPostsData } };
+};
+
+export default Home;
