+++ pages/posts/[id].tsx
@@ -1,0 +1,44 @@
+import { GetStaticPaths, GetStaticProps } from "next";
+import Head from "next/head";
+import Date from "../../components/date";
+import Layout from "../../components/layout";
+import { getAllPostIds, getPostData } from "../../lib/posts";
+import utilStyles from "../../styles/utils.module.css";
+
+interface Props {
+  postData: {
+    title: string;
+    date: string;
+    contentHtml: string;
+  };
+}
+
+const Post: React.FC<Props> = ({ postData }) => (
+  <Layout>
+    <Head>
+      <title>{postData.title}</title>
+    </Head>
+
+    <article>
+      <h1 className={utilStyles.headingXl}>{postData.title}</h1>
+
+      <div className={utilStyles.lightText}>
+        <Date dateString={postData.date} />
+      </div>
+
+      <div dangerouslySetInnerHTML={{ __html: postData.contentHtml }} />
+    </article>
+  </Layout>
+);
+
+export const getStaticPaths: GetStaticPaths = async () => {
+  const paths = getAllPostIds();
+  return { paths, fallback: false };
+};
+
+export const getStaticProps: GetStaticProps = async ({ params }) => {
+  const postData = await getPostData(params.id);
+  return { props: { postData } };
+};
+
+export default Post;
